Update: Separated the Homepage and loader
