+++ app/globals.css
@@ -63,3 +63,4 @@ body {
 .animate-fadeIn {
   animation: fadeIn 1s ease-out forwards;
 }
+
