+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { Bebas_Neue, Montserrat, Titillium_Web } from "next/font/google";
 import "./globals.css";
 // import Navbar from "@/components/Navbar"; // Removed from here
 // import Footer from "@/components/Footer"; // REMOVED: Must be rendered conditionally in page.tsx
-import LogoIntro from "@/components/LogoIntro";
+import Loader from "@/components/Loader";
 
 // 1. Headline Font: Bebas Neue
 const bebasNeue = Bebas_Neue({
@@ -43,7 +43,7 @@ export default function RootLayout({
         className={`${bebasNeue.variable} ${montserrat.variable} ${titillium.variable} antialiased bg-[#050505] text-white overflow-x-hidden`}
       >
         {/* LogoIntro is rendered first. It is fixed and hides the entire screen until the animation is complete. */}
-        <LogoIntro />
+        <Loader />
         
         {/* Main Content Area: Now completely reliant on children (page.tsx) */}
         <div className="relative z-0">
